rename test and minor change on account, change directory of accounts class
