+++ src/test/java/com/example/td_banque_java/model/integration/AccountIntegrationTest.java
@@ -1,11 +1,5 @@
 package com.example.td_banque_java.model.integration;
 
-import com.example.td_banque_java.model.Account;
-import com.example.td_banque_java.model.Bank;
-import com.example.td_banque_java.model.Client;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
