+++ src/test/java/com/example/td_banque_java/model/unit/AccountTest.java
@@ -1,13 +1,4 @@
 package com.example.td_banque_java.model.unit;
 
-import com.example.td_banque_java.model.Account;
-import com.example.td_banque_java.model.Bank;
-import com.example.td_banque_java.model.Client;
-import com.example.td_banque_java.model.WithoutOverdraftAccount;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
 class AccountTest {
 }
