+++ src/test/java/com/example/td_banque_java/model/unit/BankTest.java
@@ -1,9 +1,9 @@
 package com.example.td_banque_java.model.unit;
 
-import com.example.td_banque_java.model.Account;
+import com.example.td_banque_java.model.account.Account;
 import com.example.td_banque_java.model.Bank;
 import com.example.td_banque_java.model.Client;
-import com.example.td_banque_java.model.WithoutOverdraftAccount;
+import com.example.td_banque_java.model.account.WithoutOverdraftAccount;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -41,7 +41,7 @@ class BankTest {
     @Test
     void openAccount() {
         Client newClient = new Client("2", "client2", "address2", new Bank());
-        Account newAccount = new Account(2000, "2", newClient, new Bank());
+        Account newAccount = new WithoutOverdraftAccount(2000, "2", newClient, new Bank());
         bank.openAccount(newAccount, newClient);
 
         assertTrue(bank.getAccountsByClient().containsKey(newAccount));
