+++ src/test/java/com/example/td_banque_java/model/unit/OverdrawnAccountTest.java
@@ -2,7 +2,7 @@ package com.example.td_banque_java.model.unit;
 
 import com.example.td_banque_java.model.Bank;
 import com.example.td_banque_java.model.Client;
-import com.example.td_banque_java.model.OverdrawnAccount;
+import com.example.td_banque_java.model.account.OverdrawnAccount;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -19,6 +19,42 @@ class OverdrawnAccountTest {
     }
 
     @Test
+    void debit_validAmount_Success() {
+        account.debit(100);
+        assertEquals(900, account.getBalance());
+    }
+
+    @Test
+    void debit_overdrawn_throwIllegalArgumentException() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            account.debit(2000);
+        });
+    }
+
+    @Test
+    void debit_negativeValue_throwIllegalArgumentException() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            account.debit(-200);
+        });
+    }
+
+    @Test
+    void credit_positiveValue_increaseBalanceTo() {
+        account.credit(100);
+        assertEquals(1100, account.getBalance());
+    }
+
+    @Test
+    void credit_negativeValue_throwIllegalArgumentException() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            account.credit(-100);
+        });
+    }
+
+
+    //mutators tests
+
+    @Test
     void getOverdrawnAutorized() {
         assertEquals(100, account.getOverdrawnAutorized());
     }
@@ -28,41 +64,6 @@ class OverdrawnAccountTest {
         account.setOverdrawnAutorized(200);
         assertEquals(200, account.getOverdrawnAutorized());
     }
-
-    @Test
-    void debit() throws IllegalAccessException {
-        account.debit(100);
-        assertEquals(900, account.getBalance());
-    }
-
-    @Test
-    void debitErrorAmountNegatif() {
-        assertThrows(IllegalAccessException.class, () -> {
-            account.debit(-100);
-        });
-    }
-
-    @Test
-    void debitErrorOverdrawnAutorized() {
-        account.setBalance(0);
-        assertThrows(IllegalAccessException.class, () -> {
-            account.debit(-200);
-        });
-    }
-
-    @Test
-    void credit() throws IllegalAccessException {
-        account.credit(100);
-        assertEquals(1100, account.getBalance());
-    }
-
-    @Test
-    void creditError() {
-        assertThrows(IllegalAccessException.class, () -> {
-            account.credit(-100);
-        });
-    }
-
     @Test
     void getSold() {
         assertEquals(1000, account.getBalance());
